+++ python/lsst/ts/MTAOS/ConfigByFile.py
@@ -19,10 +19,11 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+__all__ = ["ConfigByFile"]
 
 from lsst.ts.wep.ParamReader import ParamReader
 
-from lsst.ts.MTAOS.ConfigDefault import ConfigDefault
+from .ConfigDefault import ConfigDefault
 
 
 class ConfigByFile(ConfigDefault):