+++ python/lsst/ts/MTAOS/Utility.py
@@ -19,7 +19,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+__all__ = ["WEPWarning", "OFCWarning", "getModulePath", "getConfigDir",
+           "getSchemaDir", "getLogDir", "getIsrDirPath", "getCamType",
+           "getInstName", "getCscName", "addRotFileHandler"]
+
 import os
+import logging
+from logging.handlers import RotatingFileHandler
 from enum import Enum, auto
 from pathlib import Path
 from lsst.utils import getPackageDir
@@ -73,6 +79,18 @@ def getSchemaDir():
     """
 
     return getModulePath().joinpath("schema")
+
+
+def getLogDir():
+    """Get the directory of log files.
+
+    Returns
+    -------
+    pathlib.PosixPath
+        Directory of log files.
+    """
+
+    return getModulePath().joinpath("logs")
 
 
 def getIsrDirPath(isrDirPathVar="ISRDIRPATH"):
@@ -163,6 +181,42 @@ def getCscName():
     """
 
     return "MTAOS"
+
+
+def addRotFileHandler(log, filePath, maxBytes=1e6, backupCount=5):
+    """Add a rotating file handler to a logger.
+
+    Note: The input log object will be updated directly.
+
+    Parameters
+    ----------
+    log : logging.Logger
+        Logger.
+    filePath : pathlib.PosixPath
+        File path.
+    maxBytes : int, optional
+        Maximum file size in bytes for each file. (the default is 1e6.)
+    backupCount : int, optional
+        Number of log files to retain. (the default is 5.)
+
+    Returns
+    -------
+    fileHandler : logging.RotatingFileHandler
+        The file handler added.
+    """
+
+    fileHandler = RotatingFileHandler(
+        filename=filePath, mode="a", maxBytes=int(maxBytes),
+        backupCount=int(backupCount))
+
+    logFormat = logging.Formatter(fmt="%(asctime)s - %(levelname)s - %(message)s")
+    fileHandler.setFormatter(logFormat)
+
+    fileHandler.setLevel(logging.DEBUG)
+
+    log.addHandler(fileHandler)
+
+    return fileHandler
 
 
 if __name__ == "__main__":