+++ python/lsst/ts/MTAOS/__init__.py
@@ -19,7 +19,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-from .MtaosCsc import MtaosCsc
+from .Utility import *
+from .CalcTime import *
+from .ConfigDefault import *
+from .ConfigByObj import *
+from .ConfigByFile import *
+from .Model import *
+from .ModelSim import *
+from .MtaosCsc import *
 
 # The version file is gotten by the scons. However, the scons does not support
 # the build without unit tests. This is a needed function for the Jenkins to